made section package
sections are now abstracted into a package and can be custom made easily
with the Section base class.

constants also moved to be less cluttering for the main file.
should consider embedding the constants in a json file as well and parse
it with the config loader...

added small todo list for the editor.