properly exits and added enemies
also optimized the data store on the server side, and restructured the
order of a few packets handlers.

the game now closes down the client and sends the disconnect signal to
the server before quitting and closes all threads properly.

there is enemy DTO on the server and it is broadcasted to all players at a certain
interval.